+++ src/crypto/sha3/sha3.h
@@ -200,7 +200,7 @@ typedef struct {
  * Absorb and finalize an input seed byte array.
  * Should be used in conjunction with the shake128_squeezeblocks function.
  *
- * \warning Finalizes the seed state, should not be used in consecutive calls. \n
+ * \warning Finalizes the seed state, should not be used in consecutive calls.
  * State must be allocated by the caller.
  *
  * \param state The function state; must be allocated
@@ -260,7 +260,7 @@ void OQS_SHA3_shake128_inc_finalize(OQS_SHA3_shake128_inc_ctx *s_inc);
  *
  * \param output output buffer
  * \param outlen bytes of outbut buffer
- * \param s_inc
+ * \param s_inc The function state; must be initialized
  */
 void OQS_SHA3_shake128_inc_squeeze(uint8_t *output, size_t outlen, OQS_SHA3_shake128_inc_ctx *s_inc);
 
@@ -290,7 +290,7 @@ typedef struct {
  * Absorb and finalize an input seed byte array.
  * Should be used in conjunction with the shake256_squeezeblocks function.
  *
- * \warning Finalizes the seed state, should not be used in consecutive calls. \n
+ * \warning Finalizes the seed state, should not be used in consecutive calls.
  * State must be initialized (and zeroed) by the caller.
  *
  * \param state The function state; must be pre-initialized
@@ -408,7 +408,7 @@ void OQS_SHA3_cshake128_inc_finalize(OQS_SHA3_shake128_inc_ctx *state);
  *
  * \param output output buffer
  * \param outlen bytes of outbut buffer
- * \param state
+ * \param state The function state; must be initialized
  */
 void OQS_SHA3_cshake128_inc_squeeze(uint8_t *output, size_t outlen, OQS_SHA3_shake128_inc_ctx *state);
 
@@ -465,7 +465,7 @@ void OQS_SHA3_cshake256_inc_finalize(OQS_SHA3_shake256_inc_ctx *state);
  *
  * \param output output buffer
  * \param outlen bytes of outbut buffer
- * \param state
+ * \param state The function state, must be initialized
  */
 void OQS_SHA3_cshake256_inc_squeeze(uint8_t *output, size_t outlen, OQS_SHA3_shake256_inc_ctx *state);
 
@@ -508,9 +508,9 @@ void OQS_SHA3_cshake256_simple(uint8_t *output, size_t outlen, uint16_t cstm, co
  * \warning The output array length must not be zero.
  *
  * \param output0 The first output byte array
- * \param output0 The second output byte array
- * \param output0 The third output byte array
- * \param output0 The fourth output byte array
+ * \param output1 The second output byte array
+ * \param output2 The third output byte array
+ * \param output3 The fourth output byte array
  * \param outlen The number of output bytes to generate in every output array
  * \param in0 The first input seed byte array
  * \param in1 The second input seed byte array
@@ -519,10 +519,32 @@ void OQS_SHA3_cshake256_simple(uint8_t *output, size_t outlen, uint16_t cstm, co
  * \param inlen The number of seed bytes to process from every input array
  */
 void OQS_SHA3_shake128_4x(uint8_t *output0, uint8_t *output1, uint8_t *output2, uint8_t *output3, size_t outlen, const uint8_t *in0, const uint8_t *in1, const uint8_t *in2, const uint8_t *in3, size_t inlen);
+
+/**
+* \brief Seed 4 parallel cSHAKE-128 instances, and generate 4 arrays of pseudo-random output, using a "simplified" customization string.
+* Uses a vectorized (AVX2) implementation of cSHAKE-128.
+* Permutes and extracts the state to an output byte array.
+* The "simplified" customization string procedure is ad hoc but used in several NIST candidates.
+*
+* \warning This function has a counter period of 2^16.
+*
+* \param output0 The first output byte array
+* \param output1 The second output byte array
+* \param output2 The third output byte array
+* \param output3 The fourth output byte array
+* \param outlen The number of output bytes to generate in every output array
+* \param cstm0 The first 16bit customization integer
+* \param cstm1 The second 16bit customization integer
+* \param cstm2 The third 16bit customization integer
+* \param cstm3 The fourth 16bit customization integer
+* \param input The input seed byte array
+* \param inplen The number of seed bytes to process
+*/
+void OQS_SHA3_cshake128_simple4x(uint8_t *output0, uint8_t *output1, uint8_t *output2, uint8_t *output3, size_t outlen, uint16_t cstm0, uint16_t cstm1, uint16_t cstm2, uint16_t cstm3, const uint8_t *in, size_t inlen);
 #endif
 
 #if defined(__cplusplus)
 } // extern "C"
 #endif
 
-#endif
+#endif // OQS_SHA3_H
