+++ src/crypto/sha3/sha3x4.c
@@ -146,3 +146,66 @@ void OQS_SHA3_shake128_4x(unsigned char *output0, unsigned char *output1, unsign
 		}
 	}
 }
+
+/********** cSHAKE128 ***********/
+
+static void cshake128_simple_absorb4x(__m256i *s, uint16_t cstm0, uint16_t cstm1, uint16_t cstm2, uint16_t cstm3, const unsigned char *in, unsigned long long inlen) {
+	unsigned char *sep = (unsigned char *) s;
+	unsigned int i;
+
+	for (i = 0; i < 25; i++)
+		s[i] = _mm256_xor_si256(s[i], s[i]); // zero state
+
+	/* Absorb customization (domain-separation) string */
+	for (i = 0; i < 4; i++) {
+		sep[8 * i + 0] = 0x01;
+		sep[8 * i + 1] = 0xa8;
+		sep[8 * i + 2] = 0x01;
+		sep[8 * i + 3] = 0x00;
+		sep[8 * i + 4] = 0x01;
+		sep[8 * i + 5] = 16;
+	}
+	sep[6] = cstm0 & 0xff;
+	sep[7] = cstm0 >> 8;
+	sep[14] = cstm1 & 0xff;
+	sep[15] = cstm1 >> 8;
+	sep[22] = cstm2 & 0xff;
+	sep[23] = cstm2 >> 8;
+	sep[30] = cstm3 & 0xff;
+	sep[31] = cstm3 >> 8;
+
+	KeccakF1600_StatePermute4x(s);
+
+	/* Absorb input */
+	keccak_absorb4x(s, SHAKE128_RATE, in, in, in, in, inlen, 0x04);
+}
+
+/* N is assumed to be empty; S is assumed to have at most 2 characters */
+void OQS_SHA3_cshake128_simple4x(unsigned char *output0, unsigned char *output1, unsigned char *output2, unsigned char *output3, unsigned long long outlen,
+				 uint16_t cstm0, uint16_t cstm1, uint16_t cstm2, uint16_t cstm3, const unsigned char *in, unsigned long long inlen) {
+	__m256i s[25];
+	unsigned char t0[SHAKE128_RATE];
+	unsigned char t1[SHAKE128_RATE];
+	unsigned char t2[SHAKE128_RATE];
+	unsigned char t3[SHAKE128_RATE];
+	unsigned int i;
+
+	cshake128_simple_absorb4x(s, cstm0, cstm1, cstm2, cstm3, in, inlen);
+
+	/* Squeeze output */
+	keccak_squeezeblocks4x(output0, output1, output2, output3, outlen / SHAKE128_RATE, s, SHAKE128_RATE);
+	output0 += (outlen / SHAKE128_RATE) * SHAKE128_RATE;
+	output1 += (outlen / SHAKE128_RATE) * SHAKE128_RATE;
+	output2 += (outlen / SHAKE128_RATE) * SHAKE128_RATE;
+	output3 += (outlen / SHAKE128_RATE) * SHAKE128_RATE;
+
+	if (outlen % SHAKE128_RATE) {
+		keccak_squeezeblocks4x(t0, t1, t2, t3, 1, s, SHAKE128_RATE);
+		for (i = 0; i < outlen % SHAKE128_RATE; i++) {
+			output0[i] = t0[i];
+			output1[i] = t1[i];
+			output2[i] = t2[i];
+			output3[i] = t3[i];
+		}
+	}
+}
